Merged PR 31633: Fix address book sorting
## What's being changed

This PR fixes a bug in the sorting of address books.

## Why it's being changed

We were handling the address book as an object when in fact it is always an array. On top of that, we were getting a warning (probably related to PHP 7+):
```
Deprecated: Unparenthesized `a ? b : c ? d : e` is deprecated. Use either `(a ? b : c) ? d : e` or `a ? b : (c ? d : e)`
```

## How to review / test this change

- Go to the plugin options
- Sort your address books every which way

Related work items: #152497