Merged PR 31873: Update dotmailer-signup shortcode
## What's being changed

Merchants can now use a `[dotdigital-signup]` shortcode in their posts and pages in WordPress. `[dotmailer-signup]` will continue to function, but is marked as deprecated in line with the module re-brand.

The optional arguments that can be added to the shortcode for `showtitle` and `showdesc` both now work, also.

## Why it's being changed

The branding update needed to happen whilst preserving backwards compatibility.

## How to review / test this change

- Add a shortcode e.g. `[dotdigital-signup showtitle=0 showdesc=0]` in a post or page
- Confirm the title and description do not display in the post
- Check that the title and description still display in e.g. the footer on the same page

Related work items: #153651