Merged PR 31634: Plugin housekeeping
## What's being changed

A few housekeeping items. Most self-explanatory. One to pick out:

- replace the undefined `$instance` variable in dm_shortcode.php. Resolves https://wordpress.org/support/topic/undefined-variable-and-index-using-shortcode/.

## Why it's being changed

Plugin screenshots are not supposed to be in version control, either in git or SVN. They are meant to be added in an assets/ folder in the SVN root, and will work for all versions of the plugin. See https://developer.wordpress.org/plugins/wordpress-org/plugin-assets/. In any case we're refreshing them all, so they can go.

Related work items: #152154