Merged PR 33667: Remove stats harvesting
## What's being changed

This PR removes a half-baked capability from the plugin, whereby system data was collected from the WordPress instance and emailed back to a Dotdigital email address.

## Why it's being changed

This function wasn't working, not least because the email address associated had been inactive for some time. Moreover it exposed API creds in public code. Not to mention it being a very grey activity in general to be collecting data without consent.

## How to review / test this change

- Re-test the function of the form, including posting data on submit

Related work items: #159843