Merged PR 34393: Do not expose API credentials in the event of a fatal error
## What's being changed

This PR:
- modifies how API credentials are passed to the `DotMailerConnect` constructor.
- fixes a possible issue with autoloading

## Why it's being changed

Both problems were discovered via the same case, of a merchant submitting the form and receiving a PHP Fatal Error which then revealed their API creds in the error message.

The `require` for vendor/autoload.php was clearly in the wrong place; and now the creds are passed in an array, which are not exposed in the same way.

## How to review / test this change

- In the DotMailerConnect constructor change the reference to `Container` to `Chaztainer`
- On submitting the form you will receive a fatal error but your API creds should not be visible

Related work items: #162393