Merged PR 38741: Improve handling of account info checks
## What's being changed

The logic behind the plugin's 'API credentials' admin tab.

## Why it's being changed

The API call to getAccountInfo could return something other than either `false` or the expected array of values - so our checks weren't rigorous enough. It was possible for `$accountInfo` to be set and not false, but to still not have the Properties key, which would lead to several errors printed to the admin screen.

## How to review / test this change

- Enter good API creds, check validation works and that account info properties are displayed as expected.
- Enter bad API creds, check validation throws an error and that no account info properties are displayed.

Related work items: #181203