Merged PR 39744: Fix data field handling when adding contacts to address books
## What's being changed

The handling of data fields before posting to the API.

## Why it's being changed

We updated the SDK to accept an array key-value pairs for data fields, for example:
```
$r = $dotdigital->addressBooks->addContactToAddressBook(
	113946,
	'chaz@emailsim.io',
	['APPROVE_ACCOUNT' => 1, 'ACCOUNT_MANAGER' => 'dave'],
	'Single',
	'Html'
);
```

However our plugin code was putting the values into an array already, so when the request was made, it couldn't interpret a data field by the name of e.g. "0".

## How to review / test this change

- In your plugin admin, select several data fields for capture on the front end
- Also select an address book for the front end
- Sign up
- Confirm that the contact is created, in the correct address books, and has the data field values you entered

## Notes

Date-type data fields are not currently supported (story backlogged).

Related work items: #185377