Merged PR 40039: Add error handling for incorrect API credentials
## What's being changed

We've added some exception handling when fetching address books and data fields in the plugin admin.

## Why it's being changed

This was a regression in v6.0.0. If you had bad creds, or your API user was locked, and you upgraded, you would get a fatal error and not be able to fix it without hacking the `wp_options` table.

## How to review / test this change

- Connect to Dotdigital with some API creds
- In Dotdigital, set the API user you've used to 'Disabled' (this simulates what happens if a merchant upgrades, but their creds are no longer good, or the user was locked, etc.)
- Refresh the admin
- Do some additional smoke tests in the admin and on the front end to check any other impacts of bad creds or inadequate error handling

Related work items: #186770