Merged PR 42689: Fix bug with labels for addressbooks and datafields
## What's being changed

In this PR we trimmed the output value of datafield and address book labels in plugin configurations from long untidy if statement logic to ternary operator.

## Why it's being changed

It looks like the current code was injecting loads of spaces in label fields. As a result, when entering custom label text, it appears not to be saving, and the front end then displays two labels (the default one plus the new custom one).

## How to review / test this change

1. Install https://wordpress.org/plugins/dotmailer-sign-up-widget/ on wordpress 6.1.1 or later
2. Activate plugin and add valid API credentials.
3. Navigate to /wp-admin/admin.php?page=dm_form_settings&tab=my_data_fields
4. Edit Contact data field labels.
5. Save.
6. Ensure the labels are visible in the admin.
7. Repeat steps 4-6 for address book labels.
7. Add a signup form widget to your storefront.
8. Ensure the updated labels appear next to your data fields and address books.

Related work items: #198615