Merged PR 42811: Fix broken sort item functionality
## What's being changed

In this PR we modified (and improved) our sorting functions that are responsible for sorting the Datafields and the Addressbooks in the plugin configurations in order to sort object instead of array items.

## Why it's being changed

Currently these functions were designed to sort array items as this was the format of the dotdigital items in the old API version. In the new dotdigital sdk library that we released and integrate recently with our plugin all API items are in object format. This has a result our sorting functionality to break.

## How to review / test this change

- Go to plugin configurations.
- Try to sort AddressBooks and DataFields
- Confirm that the sorting works (alphabetically)
- Confirm that the sorting by default is ascending

Related work items: #198614