Merged PR 45662: Fix data field input types for widget
## What's being changed

This PR repairs the HTML of the signup form widget by setting the expected input types according to each Dotdigital data field type.

| Dotdigital data field type | HTML input type |
| -- | -- |
| Text | text |
| Date | date |
| Numeric | number |
| Yes/No | radio |

We also fixed a deprecated notice for `uasort` return type in PHP 8.1+, and cleaned up the method that outputs the field input HTML.

## Why it's being changed

Previously the HTML was invalid and had e.g. `<input type='FIRSTNAME'>`.

## How to review / test this change

- Check out the related branch on the SDK
- Add a repositories path in composer.json to your local SDK
- Run `composer update` in the plugin root folder
- Test sorting for data fields
- Test sorting for address books
- Ensure you have selected a data field corresponding to each of the types from the table above
- Add these to the widget
- Check the HTML to confirm the input types match the appropriate data field type
- Test form submission and mapping of data field values into Dotdigital

## Notes

Since the previous version of this plugin, we have refactored our PHP SDK. As a result the code now respects the revised namespace (`Dotdigital\V2`) for the v2 API.

Related work items: #212887