Merged PR 48375: Plugin refactor
## What's being changed

We've refactored the entire plugin to fit a more object-oriented boilerplate. These changes are intended to be backwards-compatible and should not affect existing merchants who upgrade to the new version.

Key functional changes:
- The plugin is now "Dotdigital for WordPress" (note the SVN handle cannot change)
- Form submissions now post to a REST endpoint.
- Submissions are processed and query args are added to the origin URL (or the redirect URL, if one is set and the submission was successful) to indicate success/failure.
- "Address Books" have been renamed to "Lists" in line with changes to the Dotdigital platform.
- Tab URLs in the plugin settings have been changed.

## Why it's being changed

Our main aim was to bring this plugin up to the standard of our Dotdigital for Woocommerce plugin. For some time we'd worked on this codebase and made some small improvements, but the amount of spaghetti code in some key files made adapting it difficult. We have tried to use WordPress best practices, and separate logic from the view layer wherever possible.

## How to review / test this change

Testing for this update needs to be thorough, and there are more steps than can be listed here. Separate document to follow.

## Notes

- The plugin's main file is still dm_signup_form.php (changing this means it would deactivate at upgrade).
- No option key names have changed, so all data is preserved.

Related work items: #221822