Merged PR 48966: Preserve stored API endpoint at upgrade for R2 and R3
## What's being changed

The option key we use to store the API endpoint for the current connection.

## Why it's being changed

The plugin name is changing, but we used the old plugin name 'dotdigital-signup-form' in the option name for API endpoint i.e. 'dotdigital-signup-form_API_endpoint'. So, if you're on R2 or R3, when you upgrade, we won't find a stored value at 'dotdigital-for-wordpress_API_endpoint', therefore your connection won't work any longer.

## How to review / test this change

- Switch to an R2 or R3 account
- Confirm that your API endpoint is updated in 'dotdigital-signup-form_API_endpoint'
- Confirm fetching lists and data fields works as expected
- Revert to an R1 account
- Delete the option 'dotdigital-for-wordpress_API_endpoint'
- Confirm fetching lists and data fields works as expected

## Notes

Our option names are inconsistent, we know this, but we decided it was better for installations to not break at upgrade. Yes, we could probably write an upgrade script to move value A from option A to option B, but it's not something we're super-concerned about.

Related work items: #231362