Merged PR 49149: Show 'nobook' message only if lists were displayed but ignored
## What's being changed

The logic around displaying the 'Please select one newsletter' message. I've also made some changes so that we are catching errors in the right place. And I also fixed a small regression caused by using `esc_url` on the query string that prevented displaying error messages next to widgets.

## Why it's being changed

It's fine to submit the form without lists, but only if there were no lists set to visible in the admin.

## How to review / test this change

- Set up a form with two visible lists
- Submit the form without checking any list > see the error message
- Submit the form with at least one list checked > no error message
- Set up a form with 1 hidden list
- Submit the form > no error message
- Set up a form with 1 hidden list and 1 visible list
- Submit the form without checking any list > see the error message
- Set up a form with no lists configured at all
- Submit the form > no error message

Related work items: #232302