Merged PR 49060: Submit signup forms via AJAX
## What's being changed

In this PR we converted the submission process of the signup widgets from POST to AJAX. To achive this we added an extra checkbox option to allow our users to set if the widget form will be submitted via ajax or post http request.

## Why it's being changed

The widget's behaviour should be unobtrusive, and in general such widgets in WordPress normally submit with a page refresh. It has been suggested in our reviews https://wordpress.org/plugins/dotmailer-sign-up-widget/#reviews Opens in new window or tab.

This change also improves the security of our code by only allowing form submissions to be processed via specific forms, using a WordPress nonce. And it is now possible to have multiple widgets on a single page, with each one submitting and handling data in isolation from the others.

## How to review / test this change

- Confirm that a new option "Enable AJAX submission" added in the widget options.
- Insert a widget in the frontend.
- Confirm that the form is submitted both via ajax or http post and a contact created in Dotdigital.
- Confirm success error messages
- Confirm that redirection works as expected
- Confirm that multiple forms in the same page behaves as expected after submission

Related work items: #181207