Merged PR 52024: Always exit() after wp_redirect
## What's being changed

We're adding `exit()` after anywhere we call `wp_redirect()`. We've also simplified / clarified what the `post()` callback method returns (`null`) in the signup controller.

## Why it's being changed

This is strongly advised in the [WordPress Codex](https://developer.wordpress.org/reference/functions/wp_redirect/). We are trying to fix a problem with the signup form failing to redirect after submission (either back to its origin page or onward to a different URL, if redirection is configured). The problem is that the form is submitted, but then loads /wp-json/dotdigital/v1/signup-widget, which does not redirect and only prints `true` (because the `post()` method returns bool?). Oddly this is not replicable locally.

## How to review / test this change

- Place a signup form via any method (block, legacy widget, shortcode)
- Submit the form
- Confirm that the origin page is loaded
- Set redirection and test that works as expected

Related work items: #244749