Merged PR 52425: Ensure required data fields are handled the same way in AJAX vs standard mode
## What's being changed

If data fields are required in a form, they will now have the data-attribute `data-required='1'` instead of `data-required='required'`.

## Why it's being changed

The form submission payload needs to match whether sent via AJAX or via standard POST request. In AJAX mode, if a data field was not required, we did not send the 'required' key, and as a result we would see this error:
```
PHP Notice: Undefined index: required in .. plugins/dotmailer-signup-widget/includes/rest/class-dotdigital-wordpress-signup-widget-controller.php on line 172
```

## How to review / test this change

- Test forms on different pages in AJAX mode and standard mode
- The form should have one field that is required, another that is not
- Compare the array keys and values of $datafield on L172 in AJAX and standard mode (they should match for the 'required' key)
- Confirm that required fields with no value still trigger an error (remove the `required` attribute - NOT the data-required attribute) - on the input to bypass browser-based validation

Related work items: #246830