Merged PR 58289: Remove add_submenu_page from admin hooks
## What's being changed

This PR removes the call to `add_submenu_page`. It keeps the `legacy_redirect()` method, but moves it to the callback function of the main plugin menu page.

## Why it's being changed

We ran into a UI bug with the Gravity Forms plugin, and traced it to this function call. I don't fully understand why it had this effect, but it must be something to do with either the arguments we were passing in, or the priority order we were setting (or not setting).

In any case, the adjustments here have fixed the issue. We don't actually need a submenu page for the plugin, which is why we were passing null as the parent slug.

## How to review / test this change

- Install Gravity Forms
- Add a form, and confirm you can edit the field order in the drag and drop UI
- Go to the Dotdigital for WordPress plugin settings page
- Test that the page initially redirects to &tab=about
- Test that going to the valid tabs works as expected, with the relevant tab showing as active
- Test that going to a legacy tab like my_address_forms redirects to the lists tab
- Test that going to any other non-existent tab redirects to the about tab

Related work items: #271751