Merged PR 61857: Support more than 1000 lists
## What's being changed

We have added a `do { ... } while ( ... )` loop to the `getLists` function for the admin lists table.
We have added a simple JS filter to allow the user to find the record they are looking for.
We have removed the Sortable UI JS to improve performance.
We have increased the transient cache for API lists and data fields to 15 minutes.

## Why it's being changed

We needed to be able to display more than 1000 lists on the lists edit screen. By using a `do while` loop, we circumvent the API's base restriction of 1000 records per request without having to modify the SDK V2 API.

## How to review / test this change

#### Patcher testing
- Make sure you are on the `develop` branch of the plugin.
- On the lists config screen and Data fields, order your lists using the sortable plugin. Be sure to remember the order you have applied and save.
- Checkout to this branch. Use the patcher plugin provided or deactivate/activate the plugin. "Patches only run on the update event and activate events and won't run if you manually update the plugin."
- Ensure your lists and data fields now have the previously saved order as a new value and the order matches the order saved on `develop`.

#### List testing
- Clear cache or do a hard reload of the page.
- Navigate to the Lists screen.
- Ensure the edit and select features work as expected.
- If possible, test with more than 1000 list records.
- Ensure the sort feature works as expected.
- Enable debug display in your env, insure you do not see a depreciation warning on the lists and data fields pages

#### NOTE
If have run the patch already you will need to remove the patch form the wp_options table to run it again.

Related work items: #231481, #255080